requirements.txt
requests>=2.25.0
beautifulsoup4>=4.9.0
lxml>=4.6.0
alembic>=1.8.0
psycopg2-binary>=2.9.0
redis>=5.0.0
sqlalchemy>=1.4.0
python-dotenv>=0.19.0
prometheus_client>=0.19.0
pytest>=7.0.0
requests-mock>=1.10.0
coverage>=7.0.0
pytest-cov>=4.0.0
cloudscraper>=1.2.0
playwright>=1.40.0
pytest-vcr>=1.0.0
pytest-alembic>=0.10.0